requirements.txt
discord.py
python-dotenv
pyautogui
opencv-python
pillow